refactor(apps/server): extract strict admission runtime state (D1-2)
Move the StrictServiceAdmissionRuntimeV1 struct + impl, the
SharedIssuerRuntimeConfigV1 and ProviderAdmissionHttpsTransportV1
helpers (346 lines) into admission/legacy/admission_runtime.rs,
joining the first slice's module tree. Includes the policy response
builder (response_for_policy_request) previously embedded mid-file.

Visibility touches are limited to pub(crate) on the moved types,
fields, and inherent methods (strict_admission.rs and main both
assemble/consume them across the new boundary).

unified_server.rs: 15,291 -> 14,953 lines. No behavior change.

Verified: cargo check -p runtime --all-targets plus cuckoo-oram /
standard-cashu-process-e2e / shared-issuer-process-e2e feature
combinations; cargo test -p runtime all green (155 passed).
