+++ apps/server/src/bin/admission/legacy/mod.rs
@@ -1,5 +1,6 @@
 //! Legacy payment/policy admission surface (extracted from `unified_server.rs`).
 
+pub(crate) mod admission_runtime;
 pub(crate) mod cashu;
 pub(crate) mod strict_admission;
